requirements.txt
asgiref==3.2.5
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
django-environ==0.4.5
django-mptt==0.11.0
djangorestframework==3.11.0
